Orders Service enforce Order quantity to be an integer

+++ services/src/orders/model/ValueValidators.ts
@@ -13,7 +13,7 @@ export class ValueValidators {
 
   public static validSku = () => z.string().trim().min(4)
 
-  public static validQuantity = () => z.number().min(1)
+  public static validQuantity = () => z.number().int().min(1)
 
   public static validPrice = () => z.number().min(0)
